cost is now Long

+++ src/commonMain/kotlin/de/lms/gj10/CreateUi.kt
@@ -47,7 +47,7 @@ class GameUi(
             RichTextData.Style.DEFAULT.copy(font = DefaultTtfFontAsBitmap)
         )
     }
-    private fun getCostText(newCost: Int): RichTextData {
+    private fun getCostText(newCost: Long): RichTextData {
         return RichTextData.fromHTML(
             "<font color=white><b>$newCost</b></font>",
             RichTextData.Style.DEFAULT.copy(font = DefaultTtfFontAsBitmap)
@@ -66,7 +66,7 @@ class GameUi(
         updateBtnActive(buildingType == BuildingType.Extractor, btnBuildingExtractor)
         updateBtnActive(buildingType == BuildingType.Turret, btnBuildingTurret)
     }
-    fun onBuildingCostChange(buildingType: BuildingType, newCost: Int) {
+    fun onBuildingCostChange(buildingType: BuildingType, newCost: Long) {
         // Retrieve the cost TextBlock associated with the buildingType
         val costTextBlock = costTextBlocks[buildingType]
 
@@ -123,6 +123,7 @@ class GameUi(
     // UI SOUNDS
     // BG Music (MACHT STEFKA)
     // Build MUSIC
+    // Build SFX
     // SHOOT SFX
 
     private fun SContainer.generateButton(
@@ -134,7 +135,7 @@ class GameUi(
         mainImg: Bitmap,
         hotKey: Char? = null,
         type: BuildingType,
-        cost: Int,
+        cost: Long,
     ): UIButton {
         val validHotKey = hotKey?.takeIf { it in 'a'..'z' }
